Merge with latest devs fro M&M (last Summer)

+++ src/common/algo/forecasts.py
@@ -1,21 +1,24 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-#: No automatic export
-__all__ = []
+from __future__ import print_function, absolute_import, unicode_literals, division
 
 import re
 from collections import defaultdict
 
-from bronx.stdtypes.date import Time
+from bronx.stdtypes.date import Time, Month
 import footprints
-logger = footprints.loggers.getLogger(__name__)
 
 from vortex.algo.components import AlgoComponentError
 
 from vortex.util.structs import ShellEncoder
 from .ifsroot import IFSParallel
 from vortex.layout.dataflow import intent
+
+#: No automatic export
+__all__ = []
+
+logger = footprints.loggers.getLogger(__name__)
 
 
 class Forecast(IFSParallel):
@@ -34,13 +37,6 @@ class Forecast(IFSParallel):
             xpname = dict(
                 default  = 'FCST'
             ),
-            inline = dict(
-                info        = "Do inline post-processing",
-                type        = bool,
-                optional    = True,
-                default     = True,
-                doc_zorder  = -5,
-            ),
             ddhpack = dict(
                 info        = "After run, gather the DDH output file in directories.",
                 type        = bool,
@@ -50,7 +46,6 @@ class Forecast(IFSParallel):
             ),
             outputid = dict(
                 info        = "The identifier for the encoding of post-processed fields.",
-                type        = str,
                 optional    = True,
             )
         )
@@ -64,35 +59,24 @@ class Forecast(IFSParallel):
         """Default pre-link for the initial condition file"""
         super(Forecast, self).prepare(rh, opts)
 
-        sh = self.system
-
+        ininc = self.naming_convention('ic', rh)
         analysis = self.setlink(
             initrole = ('InitialCondition', 'Analysis'),
-            initname = 'ICMSH{0:s}INIT'.format(self.xpname)
+            initname = ininc()
         )
 
         if analysis:
             analysis  = analysis.pop()
             thismonth = analysis.rh.resource.date.month
 
-            def checkmonth(actualrh):
-                return bool(actualrh.resource.month == thismonth)
+            # Possibly fix the model clim
+            if self.do_climfile_fixer(rh, convkind='modelclim'):
+                self.climfile_fixer(rh, convkind='modelclim', month=thismonth,
+                                    inputrole=('GlobalClim', 'InitialClim'),
+                                    inputkind='clim_model')
 
-            self.setlink(
-                initrole = ('GlobalClim', 'InitialClim'),
-                initkind = 'clim_model',
-                initname = 'Const.Clim',
-                inittest = checkmonth,
-            )
-
-            for bdaprh in [x.rh for x in self.context.sequence.effective_inputs(
-                role = ('LocalClim', 'TargetClim', 'BDAPClim'),
-                kind = 'clim_bdap',
-            ) if x.rh.resource.month == thismonth]:
-                thisclim = bdaprh.container.localpath()
-                thisname = 'const.clim.' + bdaprh.resource.geometry.area
-                if thisclim != thisname:
-                    sh.symlink(thisclim, thisname)
+            # Possibly fix post-processing clim files
+            self.all_localclim_fixer(rh, thismonth)
 
             for iaurh in [x for x in
                           self.context.sequence.effective_inputs(role = re.compile(r'IAU_\w'))]:
@@ -108,8 +92,6 @@ class Forecast(IFSParallel):
             try:
                 namlocal = namrh.container.actualpath()
                 namc = namrh.contents
-                namc['NAMCT0'].NFPOS = int(self.inline)
-                logger.info("Setup NAMCT0's NFPOS=%d in %s", int(self.inline), namlocal)
                 if self.outputid is not None:
                     namc.setmacro('OUTPUTID', self.outputid)
                     logger.info('Setup macro OUTPUTID=%s in %s', self.outputid, namlocal)
@@ -143,7 +125,7 @@ class Forecast(IFSParallel):
                 match_pf = re_pf.match(fname)
                 if match_pf:
                     gp_map[match_pf.group(1).lower()].append(Time(match_pf.group(2)))
-            for k, v in gp_map.iteritems():
+            for k, v in gp_map.items():
                 v.sort()
                 logger.info('Gridpoint files found: domain=%s, terms=%s',
                             k,
@@ -155,7 +137,7 @@ class Forecast(IFSParallel):
         # Gather DDH in folders
         if self.ddhpack:
             ddhmap = dict(DL='dlimited', GL='global', ZO='zonal')
-            for (prefix, ddhkind) in ddhmap.iteritems():
+            for (prefix, ddhkind) in ddhmap.items():
                 flist = sh.glob('DHF{}{}+*'.format(prefix, self.xpname))
                 if flist:
                     dest = 'ddhpack_{}'.format(ddhkind)
@@ -228,7 +210,8 @@ class LAMForecast(Forecast):
 
         for i, bound in enumerate(cplrh):
             thisbound = bound.container.localpath()
-            sh.softlink(thisbound, 'ELSCF{0:s}ALBC{1:03d}'.format(self.xpname, i))
+            lbcnc = self.naming_convention('lbc', rh, actualfmt=bound.container.actualfmt)
+            sh.softlink(thisbound, lbcnc(number=i))
             if self.mksync:
                 thistool = self.synctool + '.{0:03d}'.format(i)
                 bound.mkgetpr(pr_getter=thistool, tplfetch=self.synctpl)
@@ -266,9 +249,10 @@ class DFIForecast(LAMForecast):
     def prepare(self, rh, opts):
         """Pre-link boundary conditions as special DFI files."""
         super(DFIForecast, self).prepare(rh, opts)
-        initname = 'ICMSH{0:s}INIT'.format(self.xpname)
+        ininc = self.naming_convention('ic', rh)
+        lbcnc = self.naming_convention('lbc', rh, actualfmt='fa')
         for pseudoterm in (999, 0, 1):
-            self.system.softlink(initname, 'ELSCF{0:s}ALBC{1:03d}'.format(self.xpname, pseudoterm))
+            self.system.softlink(ininc(), lbcnc(number=pseudoterm))
 
 
 class FullPos(IFSParallel):
@@ -326,7 +310,10 @@ class FullPosGeo(FullPos):
 
         # is there one (deterministic forecast) or many (ensemble forecast) fullpos to perform ?
         isMany = len(initrh) > 1
-        infile = 'ICMSH{0:s}INIT'.format(self.xpname)
+        do_fix_input_clim = self.do_climfile_fixer(rh, convkind='modelclim')
+        do_fix_output_clim = self.do_climfile_fixer(rh, convkind='targetclim', area='000')
+        ininc = self.naming_convention('ic', rh)
+        infile = ininc()
 
         for num, r in enumerate(initrh):
             str_subtitle = 'Fullpos execution on {}'.format(r.container.localpath())
@@ -338,6 +325,20 @@ class FullPosGeo(FullPos):
                     logger.critical('Cannot process multiple Historic files if %s exists.', infile)
             else:
                 sh.cp(r.container.localpath(), infile, fmt=r.container.actualfmt, intent=intent.IN)
+
+            # Fix links for climatology files
+            actualmonth = Month(r.resource.date + r.resource.term)
+            startingclim = r.resource.geometry
+
+            if do_fix_input_clim:
+                self.climfile_fixer(rh, convkind='modelclim', month=actualmonth, geo=startingclim,
+                                    inputrole=(re.compile('^Clim'), re.compile('Clim$')),
+                                    inputkind='clim_model')
+
+            if do_fix_output_clim:
+                self.climfile_fixer(rh, convkind='targetclim', month=actualmonth, notgeo=startingclim,
+                                    inputrole=(re.compile('^Clim'), re.compile('Clim$')),
+                                    inputkind='clim_model', area='000')
 
             # Standard execution
             super(FullPosGeo, self).execute(rh, opts)
@@ -392,7 +393,6 @@ class FullPosBDAP(FullPos):
             ),
             outputid = dict(
                 info        = "The identifier for the encoding of post-processed fields.",
-                type        = str,
                 optional    = True,
             ),
             server_run = dict(
@@ -427,22 +427,25 @@ class FullPosBDAP(FullPos):
         )]
         initrh.sort(key=lambda rh: rh.resource.term)
 
+        do_fix_input_clim = self.do_climfile_fixer(rh, convkind='modelclim')
+
+        ininc = self.naming_convention('ic', rh)
+        infile = ininc()
+
         for r in initrh:
             sh.subtitle('Loop on {0:s}'.format(r.resource.term.fmthm))
-            thesenames = list()
 
             thisdate = r.resource.date + r.resource.term
             thismonth = thisdate.month
             logger.info('Fullpos <month:%s>' % thismonth)
-            for bdaprh in [x.rh for x in self.context.sequence.effective_inputs(
-                role = 'LocalClim',
-                kind = 'clim_bdap',
-            ) if x.rh.resource.month == thismonth]:
-                thisclim = bdaprh.container.localpath()
-                thisname = 'const.clim.' + bdaprh.resource.geometry.area
-                if thisclim != thisname:
-                    thesenames.append(thisname)
-                    sh.symlink(thisclim, thisname)
+
+            if do_fix_input_clim:
+                self.climfile_fixer(rh, convkind='modelclim',
+                                    month=thismonth, geo=r.resource.geometry,
+                                    inputrole=(re.compile('^Clim'), re.compile('Clim$')),
+                                    inputkind='clim_model')
+
+            thesenames = self.all_localclim_fixer(rh, thismonth)
 
             # Set a local storage place
             runstore = 'RUNOUT' + r.resource.term.fmtraw
@@ -463,17 +466,20 @@ class FullPosBDAP(FullPos):
                 raise
 
             # Define an selection namelist
-            try:
-                namxt = [x for x in namxx if x.resource.term == r.resource.term].pop()
-                sh.remove('xxt00000000')
-                sh.symlink(namxt.container.localpath(), 'xxt00000000')
-            except Exception:
-                logger.critical('Could not get a selection namelist for term %s', r.resource.term)
-                raise
+            if namxx:
+                namxt = [x for x in namxx if x.resource.term == r.resource.term]
+                if namxt:
+                    sh.remove('xxt00000000')
+                    sh.symlink(namxt.pop().container.localpath(), 'xxt00000000')
+                else:
+                    logger.critical('Could not get a selection namelist for term %s', r.resource.term)
+                    raise AlgoComponentError()
+            else:
+                logger.info("No selection namelist are provided.")
 
-            # Finaly set the actual init file
-            sh.remove('ICMSH{0:s}INIT'.format(self.xpname))
-            sh.softlink(r.container.localpath(), 'ICMSH{0:s}INIT'.format(self.xpname))
+            # Finally set the actual init file
+            sh.remove(infile)
+            sh.softlink(r.container.localpath(), infile)
 
             # Standard execution
             super(FullPosBDAP, self).execute(rh, opts)
